% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towerWaterUse-package.R
\docType{package}
\name{towerWaterUse-package}
\alias{towerWaterUse-package}
\title{Get Water Use from Tower Cooling units at thermoelectric plants}
\description{
\code{towerWaterUse} put a description here
}
\details{
Leave this section alone.
\tabular{ll}{
Package: \tab towerWaterUse\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}

\tabular{ll}{
\code{\link{myDataObject}} \tab saved and documented data object. Link will open 
documentation on the data object\cr \cr

\code{\link{myFunc}} \tab Function in the package with description. Link will open 
documentation on the function\cr \cr

add more functions and data objects here using format above
}
}
\keyword{Thermoelectric}
